refactor(FON-468-1): move summary root components into components
