+++ apps/client/src/pages/summary/SummaryPage.tsx
@@ -3,11 +3,11 @@ import { useParams } from 'react-router';
 import { AuthGuard } from '@/components/guards/AuthGuard';
 import { ArchiveBannerPortal } from '@/components/shared/layouts/archived-banner/ArchiveBannerPortal';
 import { AUTHORIZED_ROLES } from '@/constants/authorized-roles.constants';
+import { Summary } from '@/features/summary/components/Summary';
 import { SummaryContainer } from '@/features/summary/components/SummaryContainer';
+import { SummaryNotFound } from '@/features/summary/components/SummaryNotFound';
 import { SummaryContext } from '@/features/summary/context/SummaryContext';
 import { useVisibleSummarySections } from '@/features/summary/hooks/useVisibleSummarySections';
-import { Summary } from '@/features/summary/Summary';
-import { SummaryNotFound } from '@/features/summary/SummaryNotFound';
 import { useIsSg } from '@/hooks/roles.hook';
 import { HttpException } from '@/utils/http-exception';
 import { useUser } from '@queries/auth.queries';
